The type of the variable in the Developer class and in the methods that work with it has been changed

+++ src/main/java/babinvas/herokutrialrun/entities/Developer.java
@@ -5,22 +5,22 @@ import java.util.List;
 
 public class Developer {
 	final private static List<Developer> developers = Arrays.asList(
-			new Developer(1L, "Станислав"),
-			new Developer(2L, "Михаил"),
-			new Developer(3L, "Георгий"),
-			new Developer(4L, "Crazy Bot %)"),
-			new Developer(5L, "Василий")
+			new Developer(1, "Станислав"),
+			new Developer(2, "Михаил"),
+			new Developer(3, "Георгий"),
+			new Developer(4, "Crazy Bot %)"),
+			new Developer(5, "Василий")
 	);
 
-	final private long id;
+	final private int id;
 	final private String name;
 
-	private Developer(long id, String name) {
+	private Developer(int id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 
-	public static boolean hasId(long developerId) {
+	public static boolean hasId(int developerId) {
 		for (Developer developer : developers) {
 			if (developerId == developer.id) {
 				return true;
@@ -30,7 +30,7 @@ public class Developer {
 		return false;
 	}
 
-	public static String getName(long developerId) {
+	public static String getName(int developerId) {
 		for (Developer developer : developers) {
 			if (developerId == developer.id) {
 				return developer.name;
